Move stores to specific package
